# Improve block serde for int, short, byte and int128 blocks