--- a/presto-spi/src/main/java/io/prestosql/spi/block/ByteArrayBlockBuilder.java
+++ b/presto-spi/src/main/java/io/prestosql/spi/block/ByteArrayBlockBuilder.java
@@ -13,6 +13,8 @@
  */
 package io.prestosql.spi.block;
 
+import io.airlift.slice.Slice;
+import io.airlift.slice.Slices;
 import org.openjdk.jol.info.ClassLayout;
 
 import javax.annotation.Nullable;
@@ -285,6 +287,11 @@ public class ByteArrayBlockBuilder
         sb.append("positionCount=").append(getPositionCount());
         sb.append('}');
         return sb.toString();
+    }
+
+    Slice getValuesSlice()
+    {
+        return Slices.wrappedBuffer(values, 0, positionCount);
     }
 
     private void checkReadablePosition(int position)