--- a/presto-spi/src/main/java/io/prestosql/spi/block/Int128ArrayBlock.java
+++ b/presto-spi/src/main/java/io/prestosql/spi/block/Int128ArrayBlock.java
@@ -13,6 +13,8 @@
  */
 package io.prestosql.spi.block;
 
+import io.airlift.slice.Slice;
+import io.airlift.slice.Slices;
 import org.openjdk.jol.info.ClassLayout;
 
 import javax.annotation.Nullable;
@@ -223,6 +225,11 @@ public class Int128ArrayBlock
         sb.append("positionCount=").append(getPositionCount());
         sb.append('}');
         return sb.toString();
+    }
+
+    Slice getValuesSlice()
+    {
+        return Slices.wrappedLongArray(values, positionOffset * 2, positionCount * 2);
     }
 
     private void checkReadablePosition(int position)