--- a/presto-spi/src/main/java/io/prestosql/spi/block/Int128ArrayBlockBuilder.java
+++ b/presto-spi/src/main/java/io/prestosql/spi/block/Int128ArrayBlockBuilder.java
@@ -13,6 +13,8 @@
  */
 package io.prestosql.spi.block;
 
+import io.airlift.slice.Slice;
+import io.airlift.slice.Slices;
 import org.openjdk.jol.info.ClassLayout;
 
 import javax.annotation.Nullable;
@@ -306,6 +308,11 @@ public class Int128ArrayBlockBuilder
         sb.append("positionCount=").append(getPositionCount());
         sb.append('}');
         return sb.toString();
+    }
+
+    Slice getValuesSlice()
+    {
+        return Slices.wrappedLongArray(values, 0, positionCount * 2);
     }
 
     private void checkReadablePosition(int position)