--- a/presto-spi/src/main/java/io/prestosql/spi/block/IntArrayBlock.java
+++ b/presto-spi/src/main/java/io/prestosql/spi/block/IntArrayBlock.java
@@ -13,6 +13,8 @@
  */
 package io.prestosql.spi.block;
 
+import io.airlift.slice.Slice;
+import io.airlift.slice.Slices;
 import org.openjdk.jol.info.ClassLayout;
 
 import javax.annotation.Nullable;
@@ -215,6 +217,11 @@ public class IntArrayBlock
         sb.append("positionCount=").append(getPositionCount());
         sb.append('}');
         return sb.toString();
+    }
+
+    Slice getValuesSlice()
+    {
+        return Slices.wrappedIntArray(values, arrayOffset, positionCount);
     }
 
     private void checkReadablePosition(int position)