--- a/presto-spi/src/main/java/io/prestosql/spi/block/ShortArrayBlockBuilder.java
+++ b/presto-spi/src/main/java/io/prestosql/spi/block/ShortArrayBlockBuilder.java
@@ -13,6 +13,8 @@
  */
 package io.prestosql.spi.block;
 
+import io.airlift.slice.Slice;
+import io.airlift.slice.Slices;
 import org.openjdk.jol.info.ClassLayout;
 
 import javax.annotation.Nullable;
@@ -285,6 +287,11 @@ public class ShortArrayBlockBuilder
         sb.append("positionCount=").append(getPositionCount());
         sb.append('}');
         return sb.toString();
+    }
+
+    Slice getValuesSlice()
+    {
+        return Slices.wrappedShortArray(values, 0, positionCount);
     }
 
     private void checkReadablePosition(int position)