Benchmark from https://github.com/prestosql/presto/pull/5972 results (used throughput so that numbers are bigger):

name | null chance | before | after | throughput increase
-- | -- | -- | -- | --
BenchmarkBlockSerde.deserializeByteNoNull | N/A | 5,41 | 0,143 | 3683%
BenchmarkBlockSerde.deserializeByteWithNull | 0,01 | 6,35 | 2,978 | 113%
BenchmarkBlockSerde.deserializeByteWithNull | 0,1 | 6,437 | 3,707 | 74%
BenchmarkBlockSerde.deserializeByteWithNull | 0,5 | 7,994 | 2,15 | 272%
BenchmarkBlockSerde.deserializeByteWithNull | 0,9 | 2,569 | 3,509 | -27%
BenchmarkBlockSerde.deserializeByteWithNull | 0,99 | 1,325 | 2,76 | -52%
BenchmarkBlockSerde.deserializeIntegerNoNull | N/A | 9,539 | 0,618 | 1444%
BenchmarkBlockSerde.deserializeIntegerWithNull | 0,01 | 10,402 | 3,188 | 226%
BenchmarkBlockSerde.deserializeIntegerWithNull | 0,1 | 9,692 | 4,011 | 142%
BenchmarkBlockSerde.deserializeIntegerWithNull | 0,5 | 9,356 | 3,25 | 188%
BenchmarkBlockSerde.deserializeIntegerWithNull | 0,9 | 3,045 | 3,417 | -11%
BenchmarkBlockSerde.deserializeIntegerWithNull | 0,99 | 1,719 | 2,892 | -41%
BenchmarkBlockSerde.deserializeLineitem | N/A | 1,608 | 0,852 | 89%
BenchmarkBlockSerde.deserializeLongDecimalNoNull | N/A | 20,521 | 2,438 | 742%
BenchmarkBlockSerde.deserializeLongDecimalWithNull | 0,01 | 6,649 | 6,149 | 8%
BenchmarkBlockSerde.deserializeLongDecimalWithNull | 0,1 | 6,986 | 7,13 | -2%
BenchmarkBlockSerde.deserializeLongDecimalWithNull | 0,5 | 8,095 | 7,974 | 2%
BenchmarkBlockSerde.deserializeLongDecimalWithNull | 0,9 | 3,889 | 3,76 | 3%
BenchmarkBlockSerde.deserializeLongDecimalWithNull | 0,99 | 3,166 | 3,071 | 3%
BenchmarkBlockSerde.deserializeShortNoNull | N/A | 10,215 | 0,307 | 3227%
BenchmarkBlockSerde.deserializeShortWithNull | 0,01 | 11,052 | 3,106 | 256%
BenchmarkBlockSerde.deserializeShortWithNull | 0,1 | 10,28 | 3,894 | 164%
BenchmarkBlockSerde.deserializeShortWithNull | 0,5 | 9,403 | 2,98 | 216%
BenchmarkBlockSerde.deserializeShortWithNull | 0,9 | 2,937 | 3,495 | -16%
BenchmarkBlockSerde.deserializeShortWithNull | 0,99 | 1,581 | 3,023 | -48%
BenchmarkBlockSerde.serializeByteNoNull | N/A | 9,686 | 0,212 | 4469%
BenchmarkBlockSerde.serializeByteWithNull | 0,01 | 9,372 | 2,414 | 288%
BenchmarkBlockSerde.serializeByteWithNull | 0,1 | 9,452 | 4,187 | 126%
BenchmarkBlockSerde.serializeByteWithNull | 0,5 | 9,534 | 2,409 | 296%
BenchmarkBlockSerde.serializeByteWithNull | 0,9 | 4,66 | 2,751 | 69%
BenchmarkBlockSerde.serializeByteWithNull | 0,99 | 2,761 | 1,353 | 104%
BenchmarkBlockSerde.serializeIntegerNoNull | N/A | 9,41 | 0,82 | 1048%
BenchmarkBlockSerde.serializeIntegerWithNull | 0,01 | 9,804 | 2,78 | 253%
BenchmarkBlockSerde.serializeIntegerWithNull | 0,1 | 10,77 | 4,151 | 159%
BenchmarkBlockSerde.serializeIntegerWithNull | 0,5 | 10,836 | 2,656 | 308%
BenchmarkBlockSerde.serializeIntegerWithNull | 0,9 | 4,862 | 3,178 | 53%
BenchmarkBlockSerde.serializeIntegerWithNull | 0,99 | 3,426 | 1,917 | 79%
BenchmarkBlockSerde.serializeLineitem | N/A | 4,616 | 3,379 | 37%
BenchmarkBlockSerde.serializeLongDecimalNoNull | N/A | 13,069 | 3,21 | 307%
BenchmarkBlockSerde.serializeLongDecimalWithNull | 0,01 | 12,215 | 5,607 | 118%
BenchmarkBlockSerde.serializeLongDecimalWithNull | 0,1 | 12,333 | 6,37 | 94%
BenchmarkBlockSerde.serializeLongDecimalWithNull | 0,5 | 10,484 | 7,319 | 43%
BenchmarkBlockSerde.serializeLongDecimalWithNull | 0,9 | 5,715 | 5,507 | 4%
BenchmarkBlockSerde.serializeLongDecimalWithNull | 0,99 | 2,634 | 3,313 | -20%
BenchmarkBlockSerde.serializeShortNoNull | N/A | 8,873 | 0,414 | 2043%
BenchmarkBlockSerde.serializeShortWithNull | 0,01 | 9,499 | 2,282 | 316%
BenchmarkBlockSerde.serializeShortWithNull | 0,1 | 10,377 | 3,75 | 177%
BenchmarkBlockSerde.serializeShortWithNull | 0,5 | 11,053 | 2,399 | 361%
BenchmarkBlockSerde.serializeShortWithNull | 0,9 | 4,682 | 2,807 | 67%
BenchmarkBlockSerde.serializeShortWithNull | 0,99 | 2,927 | 1,567 | 87%

Deserializing very sparse data (a lot of nulls) is slower than before. That was probably the case on long blocks too. I have few ideas to fixed it later.