@sopel39 2020-11-20
> Deserializing very sparse data (a lot of nulls) is slower than before. That was probably the case on long blocks too. I have few ideas to fixed it later.

So this regression was introduced by mine changes? IIRC after my changes serde for longs withs null was similar to serde without nulls before my changes.