@sopel39 2020-11-20
Could you add tests like `io.prestosql.spi.block.TestLongArrayBlockEncoding` for other blocks?