@skrzypo987 2020-11-20
> So this regression was introduced by mine changes?

It appears so. But this is not that big of a deal.
Take a look at https://docs.google.com/spreadsheets/d/1kwG8jYMCGqpxJd4wxz9pltPwyNHgzv2JgHAZq6HcMB8/edit#gid=0
Th regression is only in very sparse data sets. And it is lower in bigger type sizes. I'd say there was no or very minor regression on long type. However the byte type shows some problems. Let's stand by with merging this PR until I sort it out.