@sopel39 2020-11-20
> Th regression is only in very sparse data sets. And it is lower in bigger type sizes. I'd say there was no or very minor regression on long type. However the byte type shows some problems. Let's stand by with merging this PR until I sort it out.

How could there be a regression optimized case with nulls was as fast as non-optimized case without nulls (essentially the best case)?