@martint 2020-11-20
Can you post the std dev for the benchmark results? BTW, here's a great site for analyzing results of JMH benchmarks: https://jmh.morethan.io/