@skrzypo987 2020-11-23
Full benchmark results: https://drive.google.com/file/d/17BusbXYbd8Y5BVWXvCrNbrQz369RITMH/view?usp=sharing
Errors below 1%