@skrzypo987 2020-11-23
> Could you add tests like `io.prestosql.spi.block.TestLongArrayBlockEncoding` for other blocks?

https://github.com/prestosql/presto/pull/6055