@skrzypo987 2020-11-25
Made changes to deserialize int128. It appears that using System.arraycopy is (!) more performant than copying two longs from array to array. At least in this case.
Anyway, this is the most I can get without doing disassembly. I suggest merging it as it is now (I can throw away the byte and/or short deserialization if we cannot handle that kind of regression) and then immerse into disassembly, branch-prediction, cache lines or whatever is actually going on in here.