@sopel39 2020-11-25
> Made changes to deserialize int128. It appears that using System.arraycopy is (!) more performant than copying two longs from array to array. At least in this case.

Are you sure it isn't the problem with `2*x` multiplications there? There were 4 or more such multiplications in the code