@skrzypo987 2020-11-25
> Are you sure it isn't the problem with `2*x` multiplications there?

I did the benchmark on this. There are 2 variables `position` and `i`. Making `position` decrease by 2 helped a little bit. Doing the same with `i` decreased the performance as it needed to be divided by 2 every time anyway.